Fix a few compilation errors with Scene lights

+++ ionScene/CLight.cpp
@@ -17,7 +17,7 @@ namespace ion
 			this->Position = Position;
 		}
 
-		uint CLight::GetAttributeCount() const
+		size_t CLight::GetAttributeCount() const
 		{
 			return Attributes.size();
 		}
